orders feature + signalr integration

+++ client/src/app/features/checkout/checkout.component.ts
@@ -4,7 +4,7 @@ import { MatStepper, MatStepperModule } from '@angular/material/stepper';
 import { Router, RouterLink } from '@angular/router';
 import { MatButtonModule } from '@angular/material/button';
 import { StripeService } from '../../core/services/stripe.service';
-import { ConfirmationToken, StripeAddressElement, StripeAddressElementChangeEvent, StripePaymentElement, StripePaymentElementChangeEvent } from '@stripe/stripe-js';
+import { ConfirmationToken,StripeAddressElement, StripeAddressElementChangeEvent, StripePaymentElement, StripePaymentElementChangeEvent } from '@stripe/stripe-js';
 import { SnackbarService } from '../../core/services/snackbar.service';
 import {MatCheckboxChange, MatCheckboxModule} from '@angular/material/checkbox';
 import { StepperSelectionEvent } from '@angular/cdk/stepper';
@@ -17,6 +17,8 @@ import { Cart } from '../../shared/models/cart';
 import { CartService } from '../../core/services/cart.service';
 import { CurrencyPipe, JsonPipe } from '@angular/common';
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
+import { OrderToCreate, ShippingAddress } from '../../shared/models/order';
+import { OrderService } from '../../core/services/order.service';
 
 @Component({
   selector: 'app-checkout',
@@ -40,6 +42,7 @@ export class CheckoutComponent implements OnInit, OnDestroy {
   private snackbar = inject(SnackbarService);
   private router = inject(Router);
   private accountService = inject(AccountService);
+  private orderService = inject(OrderService);
   cartService = inject(CartService);
   addressElement?: StripeAddressElement;
   paymentElement?: StripePaymentElement;
@@ -100,7 +103,7 @@ export class CheckoutComponent implements OnInit, OnDestroy {
   async onStepChange(event: StepperSelectionEvent) {
     if (event.selectedIndex === 1 ) {
       if (this.saveAddress) {
-        const address = await this. getAddressFromStripedAddress();
+        const address = await this. getAddressFromStripedAddress() as Address;
         address && firstValueFrom (this.accountService.updateAddress(address));
     }
     }
@@ -118,15 +121,26 @@ export class CheckoutComponent implements OnInit, OnDestroy {
     try {
       if(this.confirmationToken){
         const result = await this.stripeService.confirmPayment(this.confirmationToken);
-        if(result.error) { 
-          throw new Error(result.error.message);
-        } else {
-        this.cartService.deleteCart();
-        this.cartService.selectedDelivery.set(null);
-       this.router.navigateByUrl('/checkout/success');
 
-        }
-      }
+        if (result.paymentIntent?.status !== 'succeeded') {
+          const order = await this.createOrderModel();
+          const orderResult = await firstValueFrom(this.orderService.createOrder(order));
+          if(orderResult) {
+         this.cartService.deleteCart();
+         this.cartService.selectedDelivery.set(null);
+         this.router.navigateByUrl('/checkout/success');
+    }else {
+      throw new Error('Order creation failed after payment confirmation.');
+    }
+
+  }else if(result.error) {
+   throw new Error(result.error.message);
+
+  }else {
+    throw new Error('something went wrong during payment confirmation.');
+  }
+
+  }
     } catch (error: any) {
       this.snackbar.error(error.message || ' something went wrong during payment confirmation.');
       stepper.previous();
@@ -136,13 +150,36 @@ export class CheckoutComponent implements OnInit, OnDestroy {
 
   }
 
+  private async createOrderModel(): Promise<OrderToCreate> {
+    const cart = this.cartService.cart();
+    const shippingAddress = await this.getAddressFromStripedAddress() as ShippingAddress;
+    const card = this.confirmationToken?.payment_method_preview.card;
 
-    private async getAddressFromStripedAddress(): Promise<Address | null> {
+    if(!cart?.id || !cart.deliveryMethodId || !card || !shippingAddress){
+      throw new Error('Missing required information to create order');
+   }
+
+   return {
+      cartId : cart.id,
+      paymentSummary: {
+        last4: +card.last4,
+        brand: card.brand,
+        expMonth: card.exp_month,
+        expYear: card.exp_year
+    },
+    deliveryMethodId: cart.deliveryMethodId,
+    shippingAddress: shippingAddress
+    }
+  }
+
+
+    private async getAddressFromStripedAddress(): Promise<Address|ShippingAddress | null> {
     const result = await this.addressElement?.getValue();
     const address = result?.value.address;
 
     if(address){
       return{
+        name:result.value.name,
         line1: address.line1,
         line2: address.line2 || undefined,
         city: address.city,
